add server refactor to model and WIP index
