+++ server/src/model.js
@@ -1,3 +1,10 @@
+interface Sequelize {
+  id: number,
+  username: string,
+  wpm: number
+  tableName: 
+}
+
 module.exports = (sequelize, DataTypes) => {
   const User = sequelize.define('User', {
     id: {
